web/requirements.txt
django==2.0.12
psycopg2==2.6.1
gunicorn==19.5.0
djangorestframework==3.9.1
drf-nested-routers==0.90.0
djangorestframework-jwt==1.11.0
sentry-sdk==0.9.0
